+++ client/src/App.vue
@@ -1,32 +1,35 @@
 <template>
   <div id="app">
-    <div id="nav">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link> |
-      <router-link to="/map">Map</router-link>
-    </div>
+    <v-toolbar>
+      <v-toolbar-title>
+        PDX METRO FOOD ENVIRONMENT
+      </v-toolbar-title>
+      <v-spacer></v-spacer>
+      <div id="nav">
+        <router-link to="/">Map</router-link> |
+        <router-link to="/about">About</router-link>
+      </div>
+    </v-toolbar>
     <router-view />
   </div>
 </template>
 
 <style>
+/* GLOBAL STYLES  */
 #app {
   font-family: "Avenir", Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
-  color: #2c3e50;
-}
-#nav {
-  padding: 30px;
 }
 
-#nav a {
-  font-weight: bold;
-  color: #2c3e50;
+ul {
+  list-style-type: none;
+  padding: 0;
 }
 
-#nav a.router-link-exact-active {
-  color: #42b983;
+li {
+  display: inline-block;
+  margin: 0 10px;
 }
 </style>
