+++ client/src/router.js
@@ -20,11 +20,6 @@ export default new Router({
       // which is lazy-loaded when the route is visited.
       component: () =>
         import(/* webpackChunkName: "about" */ "./views/About.vue")
-    },
-    {
-      path: "/map",
-      name: "map",
-      component: () => import(/* webpackChunkName: "map" */ "./views/Map.vue")
     }
   ]
 });
