+++ client/src/views/About.vue
@@ -1,5 +1,7 @@
 <template>
-  <div class="about">
-    <h1>This is an about page</h1>
-  </div>
+  <v-container>
+    <v-layout>
+      <h1>About the PDX Metro Food Environment Map</h1>
+    </v-layout>
+  </v-container>
 </template>
